+++ src/components/Menus/Sidebar/SidebarPanel.jsx
@@ -11,16 +11,19 @@ const SidebarPanel = (props) => {
     heading,
     Footer,
     className,
-    buttons = [],
-    defaultWidth = "225px",
+    buttons,
+    defaultWidth = "320px",
   } = props;
   const isMobile = useMediaQuery({ query: MOBILE_MEDIA_QUERY });
+
+  const buttonsIsEmptyArray =
+    buttons && Array.isArray(buttons) && buttons.length === 0;
 
   const panelContent = (
     <>
       <div className="sidebar__panel-header">
         <h2 className="sidebar__panel-heading">{heading}</h2>
-        {buttons?.length > 0 && (
+        {buttons && !buttonsIsEmptyArray && (
           <div className="sidebar__panel-buttons">{buttons}</div>
         )}
       </div>
@@ -51,7 +54,7 @@ const SidebarPanel = (props) => {
       defaultWidth={defaultWidth}
       defaultHeight="100%"
       handleDirection="right"
-      minWidth="180px"
+      minWidth="280px"
       maxWidth="600px"
     >
       {panelContent}